Updated Stone Tower Temple
After volume issues with assigning the Choir to the channel that was "wind", there is now "ambiance" in the song that shouldn't have any silence. Unfortunately I discovered an issue that was consistent, getting near an enemy and then leaving that enemy kills the BGM until the area is reloaded. This may have to do with the instrument set being used (0x22 End Credits 2), and I fear that the other uploads I submitted might also share this issue.