Added Clock Town Tower (Majora's Mask)
uses 0x09 (Fairy Fountain). Similar to Last Hours, sounds similar and evokes the mood of the source song, however it's obviously not a 1 to 1 replica. Clock Town Tower is where you meet the Happy Mask Salesman for the first time in MM, and where you learn the Song of Healing (the basis for the melody of the song)