E-Ada93 - Added Boss Music (MM) (#12)
* Created Fanfare subfolders w/ meta file

* Added most of the Fanfares unique to Majoras Mask

If shuffling Fanfares, these will work without issue

* Added 2 additional Fanfares

Created the Ocarina theme from Link to the Past and the Recorder from Zelda 1 as fanfares. Used seq64's midi import feature and touched them up slightly.

* Updated Readme.MD

added self credit indicating what was added

* Add files via upload

* Delete LttP_Ocarina.meta

* Delete LttP_Ocarina.seq

* Flawless ports from MM

All of these sequences use instrument sets that are identical in MM (all of them use Termina Field, which is identical to Hyrule Field, except for Music Box House, which is identical to Kakariko (Guitar))

* Mostly perfect ports from MM

aside from some instrument replacements, these are roughly identical to their MM originals. Ikana Castle is listed as "2" due to the MMR remix already being present

* Update README.md

* Updated Stone Tower Temple

After volume issues with assigning the Choir to the channel that was "wind", there is now "ambiance" in the song that shouldn't have any silence. Unfortunately I discovered an issue that was consistent, getting near an enemy and then leaving that enemy kills the BGM until the area is reloaded. This may have to do with the instrument set being used (0x22 End Credits 2), and I fear that the other uploads I submitted might also share this issue.

* LttP Ocarina added in Fanfares

I couldn't find a way to lower the pitch (seq64 has a "pitch bend" function, but from what I can tell, it only raises the pitch) so I instead lowered the volume by 16 (Master Volume set to 48 instead of 58)

* Added Deku Palace (Majora's Mask)

Uses 0x22 (kill music potential). Uses somewhat high-ish bells and jingles. Original melody is recognizable but sounds vastly different compared to the source song.

* Added Great Bay (Majora's Mask)

uses 0x1B (Gerudo Valley). Attempted to get as close to the source as possible, however it might sound too close to Southern Swamp and might get confused for that.

* Added Great Bay Temple (Majora's Mask)

uses 0x15 (Water Temple). Missing major amounts of ambiance, highly incompatible source instruments made this a difficult conversion. still sounds somewhat acceptable.

* Added Last Hours (Majora's Mask)

uses 0x13 (Goddess Cutscene). It sounds a bit quiet and while it doesn't sound like the source, it still captures the mood of the source song.

* Added Milk Bar (Majora's Mask)

uses 0x08 (kakariko guitar). uses close equivalent instruments due to not having a close equivalent in OoT, sounds good just odd compared to the source.

* Added VS Majora's Mask (final boss phase 1)

uses 0x0F (kokiri forest). doesn't sound too close to the source song

* moved Vs Majora.meta

accidently uploaded at root

* deleting to move to proper location

* moved Vs Majora.seq

* Added Woodfall Temple (Majora's Mask)

uses 0x18 (Shadow Temple). Gets close to the source's mood and intention, however making due with OoT's closest equivalent might be annoying to listen to

* Added Clock Town Tower (Majora's Mask)

uses 0x09 (Fairy Fountain). Similar to Last Hours, sounds similar and evokes the mood of the source song, however it's obviously not a 1 to 1 replica. Clock Town Tower is where you meet the Happy Mask Salesman for the first time in MM, and where you learn the Song of Healing (the basis for the melody of the song)

* Delete Elegy_of_Emptiness.meta

* Delete Elegy_of_Emptiness.seq

* Delete Get_Mask.meta

* Delete Get_Mask.seq

* Delete Goron_Lullaby.meta

* Delete Goron_Lullaby.seq

* Delete New_Wave_Bossa_Nova.meta

* Delete New_Wave_Bossa_Nova.seq

* Delete Oath_to_Order.meta

* Delete Oath_to_Order.seq

* Delete Sonata_of_Awakening.meta

* Delete Sonata_of_Awakening.seq

* Delete Song_of_Double_Time.meta

* Delete Song_of_Double_Time.seq

* Delete Song_of_Healing.meta

* Delete Song_of_Healing.seq

* Delete Song_of_Inverted_Time.meta

* Delete Song_of_Inverted_Time.seq

* Delete Song_of_Soaring.meta

* Delete Song_of_Soaring.seq

* Delete Z1_Recorder.seq

* Delete Z1_Recorder.meta

* Delete LttP_Ocarina.seq

* Delete LttP_Ocarina.meta

* Delete Chase_Theme.meta

* Delete Chase_Theme.seq

* Delete Clock_Town_Day_1.meta

* Delete Clock_Town_Day_1.seq

* Delete Clock_Town_Tower.meta

* Delete Clock_Town_Tower.seq

* Delete Deku_Palace.meta

* Delete Deku_Palace.seq

* Delete Gerudo_Fortress.meta

* Delete Gerudo_Fortress.seq

* Delete Great_Bay.meta

* Delete Great_Bay.seq

* Delete Great_Bay_Temple.meta

* Delete Great_Bay_Temple.seq

* Delete Ikana_Castle2.meta

* Delete Ikana_Castle2.seq

* Delete Inverted_Stone_Tower_Temple.meta

* Delete Inverted_Stone_Tower_Temple.seq

* Delete Last_Hours.meta

* Delete Last_Hours.seq

* Delete Milk_Bar.meta

* Delete Milk_Bar.seq

* Delete Miniboss.meta

* Delete Miniboss.seq

* Delete Music_Box_House.meta

* Delete Music_Box_House.seq

* Delete Song_of_Death.meta

* Delete Song_of_Death.seq

* Delete Southern_Swamp.meta

* Delete Southern_Swamp.seq

* Delete Stone_Tower_Temple.meta

* Delete Stone_Tower_Temple.seq

* Delete Termina_Field.meta

* Delete Termina_Field.seq

* Delete VS_Majoras_Mask.meta

* Delete VS_Majoras_Mask.seq

* Delete Woodfall_Temple.meta

* Delete Woodfall_Temple.seq

* Delete Enemy.meta

* Delete Enemy.seq

* Re-added Majora's Mask sequences

properly formatted to be in line with the main branch

* Added LttP Ocarina fanfare

* Added Fanfare - Recorder

This is the tune that plays when you use the Recorder item (also famously known as the Warp Whistle in SMB3)

* Create Groose Theme.meta

* Add files via upload

* Added Boss music

* Update Final Boss Phase 3.meta

* Update Final Boss Phase 3.meta

* Update Boss Fight.meta

updating meta file to be consistent with pr #13