Ocarina of Time sequences (fanfares)
OoT Rando recently had a PR commited (1055) that allows the user to randomize the music while omitting the songs present in the game. This is purely for the purposes of customization if a user wants some Ocarina songs, but not all of them as a work around. This batch includes all of the fanfares.