Ocarina of Time sequences (bgm)
OoT Rando recently had a PR commited (1055) that allows the user to randomize the music while omitting the songs present in the game. This is purely for the purposes of customization if a user wants some Ocarina songs, but not all of them as a work around. This batch includes the songs of the game, excluding Hyrule Field (due to it taking up 22 sequence slots in the vanilla game). NOTE: Fire Temple is from v1.2 and is not optimized for v1.0 due to using Instruments that were added to the Fire Temple Instrument Set (1.0 uses 0,1,2 for Islamic Chant, 1.2 uses A,B,C,D for Choir)